add variables when applicable
Some users requested the ability to be able to include the graphql
request variables in the log line so that they could reproduce bugs more
easily. While this shouldn't be the encouraged behavior for security
purposes, we don't want to exclude those users from being able to use
the library.

This adds that ability by passing in the `include_variables: true` flag
to the plug call site.

Additionally it adds a test to ensure that getting the GraphQL operation
name and type happens as expected.

Finally it also pattern matches on the HTTP request method to only check
the graphql parameters for POST requests, as these are the only requests
supported by the GraphQL standard.