fix include_variables bug
When the `include_variables` flag was passed to the plug, any time a
post request was received without variables in the request body, an
error would occur.

I have changed the logic so that variables will only be looked for if
the request has both an `operationName` and `type`. Additionally it will
now only return variables if they exist, whether the request be a
GraphQL one or not. And appropriate regression tests have been added.