better handle queries with no operationName
Because operation names aren't a required field in GraphQL requests,
this change will properly handle queries with names in the query that
don't send an operationName parameter in the request. An operationName
parameter will be preferred though to avoid running a regular
expression.

Additionally this adds an `include_unnamed_queries` option. When
enabled, the entire query body will be logged for each GraphQL request
that doesn't have a name in either the operationName parameter or in the
query.