properly handle queries with no commas
When I originally wrote the regular expressions for extracting query
names, I was working off of the GraphQL patterns that I had seen. I just
learned that commas between variables aren't necessary according to the
spec.

This will allow for the use of variables lists without commas as well as
for whitespace inside of the parentheses.