feat: add Datadog formatter
Add a new drop in formatter, `Uinta.Formatter.Datadog` that adds the keys
`dd.trace_id` and `dd.span_id` to the Uinta JSON output. The formatter
expects the Logger metadata keys `:trace_id` and `:span_id` to be set
with the full OpenTelemetry trace id and span id in hexadecimal.

These changes are fully backward compatible with `Uinta.Formatter`