feat: Allow for Custom Status Codes to Be Sampled
I have a public web app that has an expected number of 401s/422s/etc that are not being sampled out by Uinta. This PR makes it so I can pass in a list of allowed HTTP status codes to be considered as part of the 'success' logs, and sampled according to the existing logic.